06/07

+++ src/main/java/com/example/Monitoring/controller/IncidentController.java
@@ -1,7 +1,10 @@
 package com.example.Monitoring.controller;
 
 import com.example.Monitoring.model.Incident;
+import com.example.Monitoring.model.Product;
+import com.example.Monitoring.model.User;
 import com.example.Monitoring.service.IncidentService;
+import com.example.Monitoring.service.ProductService;
 import com.example.Monitoring.service.UserService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Controller;
@@ -20,19 +23,31 @@ import java.time.LocalDateTime;
 public class IncidentController {
     private final UserService userService;
     private final IncidentService incidentService;
+    private final ProductService productService;
 
     @GetMapping("/incidents")
     public String incidents(@RequestParam(name="searchWord", required = false)Date dateSQL,
                             Principal principal, Model model){
-        if (dateSQL!=null) {
-            LocalDateTime dateTime = dateSQL.toLocalDate().atTime(12, 0);
+        User user=userService.getUserByPrincipal(principal);
+        String divisionTitle = user.getDivisionTitle();
+        if(user.isAdmin()) {
+            LocalDateTime dateTime;
+            if (dateSQL != null) {
+                dateTime = dateSQL.toLocalDate().atTime(12, 0);
+            } else {
+                dateTime = null;
+            }
             model.addAttribute("incidents", incidentService.list(dateTime));
         }else {
-            LocalDateTime dateTime =null;
-            model.addAttribute("incidents", incidentService.list(dateTime));
+            LocalDateTime dateTime;
+            if (dateSQL != null) {
+                dateTime = dateSQL.toLocalDate().atTime(12, 0);
+            } else {
+                dateTime = null;
+            }
+            model.addAttribute("incidents", incidentService.listDivision(dateTime, divisionTitle));
         }
-        model.addAttribute("user", userService.getUserByPrincipal(principal));
-        model.addAttribute("searchWord", dateSQL);
+        model.addAttribute("user", user);
         return "incidents";
     }
 
@@ -42,16 +57,27 @@ public class IncidentController {
                       @RequestParam ("description") String description,
                       @RequestParam ("dateOfIncident") Date dateOfIncident,
                       @RequestParam("industrial") String industrial,
-                      @RequestParam("document") String document){
+                      @RequestParam("document") String document,
+                      Principal principal){
         Incident incident = new Incident();
-        if(isGarant==null) isGarant=false;
-        incidentService.correct(incident, product, isGarant, description, dateOfIncident.toLocalDate().atTime(12,0),
-                industrial, document);
+        User user = userService.getUserByPrincipal(principal);
+        Product productEntity=productService.findById(product);
+        System.out.println(productEntity.getId());
+        if(isGarant==null) {
+            isGarant = false;
+        }else {
+            isGarant = true;
+        }
+        if (productEntity.getRegiment().getDivision().getTitle().equals(user.getDivisionTitle()) || user.isAdmin()) {
+            incidentService.correct(incident, product, isGarant, description, dateOfIncident.toLocalDate().atTime(12, 0),
+                    industrial, document);
+        }
         return "redirect:/incidents";
     }
 
     @GetMapping("incidents/edit/{id}")
     public String edit(@PathVariable("id") Long id, Model model, Principal principal){
+
         Incident incident = incidentService.findById(id);
         model.addAttribute("incident", incident);
         model.addAttribute("user", userService.getUserByPrincipal(principal));
@@ -60,21 +86,35 @@ public class IncidentController {
 
     @PostMapping("incidents/edit/incident-edit/{id}")
     public String edit(@RequestParam ("product") Long product,
-                       @RequestParam ("isGarant") Boolean isGarant,
+                       @RequestParam (name="isGarant", required = false) Boolean isGarant,
                        @RequestParam ("description") String description,
                        @RequestParam ("dateOfIncident") Date dateOfIncident,
                        @RequestParam("industrial") String industrial,
                        @RequestParam("document") String document,
-                       @PathVariable("id") Long id){
-        Incident incident=incidentService.findById(id);
-        incidentService.correct(incident, product, isGarant, description, dateOfIncident.toLocalDate().atTime(12,0),
-                industrial, document);
+                       @PathVariable("id") Long id,
+                       Principal principal){
+        if(isGarant==null) {
+            isGarant = false;
+        }else {
+            isGarant = true;
+        }
+        User user = userService.getUserByPrincipal(principal);
+        Product productEntity=productService.findById(product);
+        if (productEntity.getRegiment().getDivision().getTitle().equals(user.getDivisionTitle()) || user.isAdmin()) {
+            Incident incident = incidentService.findById(id);
+            incidentService.correct(incident, product, isGarant, description, dateOfIncident.toLocalDate().atTime(12, 0),
+                    industrial, document);
+        }
         return "redirect:/incidents";
     }
 
     @GetMapping("incidents/delete/{id}")
-    public String delete(@PathVariable("id") Long id){
-        incidentService.delete(id);
+    public String delete(@PathVariable("id") Long id, Principal principal){
+        User user = userService.getUserByPrincipal(principal);
+        Product productEntity=incidentService.findById(id).getProduct();
+        if (productEntity.getRegiment().getDivision().getTitle().equals(user.getDivisionTitle()) || user.isAdmin()) {
+            incidentService.delete(id);
+        }
         return "redirect:/incidents";
     }
 }
