+++ src/main/java/com/example/Monitoring/controller/ProductController.java
@@ -1,7 +1,12 @@
 package com.example.Monitoring.controller;
 
+import com.example.Monitoring.model.Division;
 import com.example.Monitoring.model.Product;
+import com.example.Monitoring.model.Regiment;
+import com.example.Monitoring.model.User;
+import com.example.Monitoring.service.DivisionService;
 import com.example.Monitoring.service.ProductService;
+import com.example.Monitoring.service.RegimentService;
 import com.example.Monitoring.service.UserService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Controller;
@@ -20,12 +25,19 @@ import java.time.LocalDateTime;
 public class ProductController {
     private final UserService userService;
     private final ProductService productService;
+    private final RegimentService regimentService;
 
     @GetMapping("/products")
     public String products(@RequestParam(name = "searchWord", required = false) String title,
                            Principal principal, Model model) {
-        model.addAttribute("products", productService.list(title));
-        model.addAttribute("user", userService.getUserByPrincipal(principal));
+        User user = userService.getUserByPrincipal(principal);
+        String divisionTitle = user.getDivisionTitle();
+        if(user.isAdmin()) {
+            model.addAttribute("products", productService.list(title));
+        }else {
+            model.addAttribute("products", productService.listByDivision(title, divisionTitle));
+        }
+        model.addAttribute("user", user);
         model.addAttribute("searchWord", title);
         return "products";
     }
@@ -34,18 +46,27 @@ public class ProductController {
     public String add(@RequestParam("serialNumber") String serialNumber,
                       @RequestParam("modelOfTechnical") String modelOfTechnical,
                       @RequestParam("regiment") String regiment,
-                      @RequestParam("dateOfProduction") Date dateSQL) {
-        Product product = new Product();
-        LocalDateTime date = dateSQL.toLocalDate().atTime(12, 0);
-        productService.correct(product, serialNumber, modelOfTechnical, regiment, date);
+                      @RequestParam("dateOfProduction") Date dateSQL,
+                      Principal principal) {
+        User user = userService.getUserByPrincipal(principal);
+        Regiment regimentEntity = regimentService.findByTitle(regiment);
+        if (regimentEntity.getDivision().getTitle().equals(user.getDivisionTitle()) || user.isAdmin()) {
+            Product product = new Product();
+            LocalDateTime date = dateSQL.toLocalDate().atTime(12, 0);
+            productService.correct(product, serialNumber, modelOfTechnical, regiment, date);
+        }
         return "redirect:/products";
     }
 
     @GetMapping("products/edit/{id}")
     public String edit(@PathVariable("id") Long id, Model model, Principal principal) {
+        User user = userService.getUserByPrincipal(principal);
         Product product = productService.findById(id);
-        model.addAttribute("product", product);
-        model.addAttribute("user", userService.getUserByPrincipal(principal));
+        String divisionTitle=product.getRegiment().getDivision().getTitle();
+        if (divisionTitle.equals(user.getDivisionTitle()) || user.isAdmin()) {
+            model.addAttribute("product", product);
+            model.addAttribute("user", userService.getUserByPrincipal(principal));
+        }
         return "product-edit";
     }
 
@@ -54,16 +75,26 @@ public class ProductController {
                        @RequestParam("modelOfTechnical") String modelOfTechnical,
                        @RequestParam("regiment") String regiment,
                        @RequestParam("dateOfProduction") Date dateSQL,
-                       @PathVariable("id") Long id) {
+                       @PathVariable("id") Long id,
+                       Principal principal) {
+        User user = userService.getUserByPrincipal(principal);
         Product product = productService.findById(id);
-        LocalDateTime date = dateSQL.toLocalDate().atTime(12, 0);
-        productService.correct(product, serialNumber, modelOfTechnical, regiment, date);
+        String divisionTitle=product.getRegiment().getDivision().getTitle();
+        if (divisionTitle.equals(user.getDivisionTitle()) || user.isAdmin()) {
+            LocalDateTime date = dateSQL.toLocalDate().atTime(12, 0);
+            productService.correct(product, serialNumber, modelOfTechnical, regiment, date);
+        }
         return "redirect:/products";
     }
 
     @GetMapping("product/delete/{id}")
-    public String delete(@PathVariable("id")Long id){
-        productService.delete(id);
+    public String delete(@PathVariable("id")Long id, Principal principal){
+        User user = userService.getUserByPrincipal(principal);
+        Product product = productService.findById(id);
+        String divisionTitle=product.getRegiment().getDivision().getTitle();
+        if (divisionTitle.equals(user.getDivisionTitle()) || user.isAdmin()) {
+            productService.delete(id);
+        }
         return "redirect:/products";
     }
